Temp and Humidity print scaled

+++ devices/dht11.c
@@ -3,7 +3,7 @@
 #define USE_DHT
 #define DHT_PARAM_TYPE          (DHT11)
 
-void dht11_init(dht11_device_t * device, gpio_t pin) {
+void dht11_init(dht11_device_t *device, gpio_t pin) {
     dht_params_t my_params;
     my_params.pin = pin;
     my_params.type = DHT_PARAM_TYPE;
@@ -12,14 +12,14 @@ void dht11_init(dht11_device_t * device, gpio_t pin) {
     dht_init(&(device->dht), &my_params);
 }
 
-void dht11_update(dht11_device_t * device) {
+void dht11_update(dht11_device_t *device) {
     dht_read(&(device->dht), &(device->last_temp), &(device->last_hum));
 }
 
-float dht11_get_temp(dht11_device_t * device) {
-    return ((float) device->last_temp) / 10.0f;
+double dht11_get_temp(dht11_device_t *device) {
+    return ((double) device->last_temp) / 10.0f;
 }
 
-float dht11_get_hum(dht11_device_t * device) {
-    return ((float) device->last_hum) / 10.0f;
+double dht11_get_hum(dht11_device_t *device) {
+    return ((double) device->last_hum) / 10.0f;
 }
